Core adjusts: Change type of rating property. Hide loading on error.

+++ core-lib/src/main/java/br/com/mxel/cuedot/data/model/IMovie.java
@@ -25,9 +25,9 @@ public interface IMovie {
 
     void setSynopsis(String synopsis);
 
-    double getRating();
+    float getRating();
 
-    void setRating(double rating);
+    void setRating(float rating);
 
     String getReleaseDate();
